can't reference read_df, trying classmethod

+++ install.py
@@ -13,17 +13,16 @@ if __name__ == '__main__':
     while None in source_packages: source_packages.remove(None)
 
     destination = Path(site.USER_SITE)
-    destination_packages = [dir for dir in os.listdir(destination)]
+    if destination.exists():
+        destination_packages = [dir for dir in os.listdir(destination)]
 
-    for package in destination_packages:
-        print(f'Delete package {package} from {destination}')
-        shutil.rmtree(os.path.join(destination, package))
+        for package in destination_packages:
+            print(f'Delete package {package} from {destination}')
+            shutil.rmtree(os.path.join(destination, package))
 
     for package in source_packages:
         print(f'Copying package {package} from {source} to {destination}')
         shutil.copytree(os.path.join(source,package), os.path.join(destination,package))
-
-
